Move PeriodSelect to UI namespace and use UI::Button
Component is now UI::PeriodSelect; stimulus controller is ui--period-select.
Period buttons use UI::ButtonLink and the Update/custom-toggle buttons use
UI::Button instead of the legacy btn-outline-secondary/btn-success classes.
