+++ app/components/admin/index_skeleton/component.html.erb
@@ -50,7 +50,7 @@
 <% end %>
 
 <div class="mt-2 mb-4">
-  <%= render PeriodSelect::Component.new(period: @period, start_time: @start_time, end_time: @end_time) %>
+  <%= render UI::PeriodSelect::Component.new(period: @period, start_time: @start_time, end_time: @end_time) %>
 </div>
 
 <% if @admin_search_form.present? %>
