--- app/components/period_select/component.en.yml
@@ -1,16 +1,0 @@
----
-en:
-  components:
-    period_select:
-      all: All
-      custom: custom
-      day: day
-      from: From
-      hour: hour
-      next: next
-      past: past
-      seven_days: 7 days
-      thirty_days: 30 days
-      to: To
-      update: Update
-      year: year
