+++ app/components/ui/period_select/component.en.yml
@@ -1,0 +1,17 @@
+---
+en:
+  components:
+    ui:
+      period_select:
+        all: All
+        custom: custom
+        day: day
+        from: From
+        hour: hour
+        next: next
+        past: past
+        seven_days: 7 days
+        thirty_days: 30 days
+        to: To
+        update: Update
+        year: year
